+++ thesocialapp/src/main/java/com/socialapp/repositories/CategoryRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.socialapp.entities.Category;
 
-public interface CategoryRepo extends CrudRepository<Category, Integer> {
+public interface CategoryRepo extends CrudRepository<Category, Long> {
 }
