+++ thesocialapp/src/main/java/com/socialapp/services/PostService.java
@@ -11,27 +11,27 @@ import org.springframework.web.multipart.commons.CommonsMultipartFile;
 public interface PostService {
 
 	// Create Post
-	PostDto createPost(PostDto postDto, Integer userId, Integer categoryId) throws IOException;
+	PostDto createPost(PostDto postDto, Long userId, Long categoryId) throws IOException;
 
 
 
     // Update Post
-	PostDto updatePost(PostDto postDto, Integer postId);
+	PostDto updatePost(PostDto postDto, Long postId);
 
 	// Delete post
-	void deletePost(Integer postId);
+	void deletePost(Long postId);
 
 	// Get all Posts
 	List<PostDto> getAllPost();
 
 	// Get Single Post
-	PostDto getPostById(Integer postid);
+	PostDto getPostById(Long postid);
 	
 	// Get all Posts by Category
-	List<PostDto> getAllPostByCategory(Integer categoryId);
+	List<PostDto> getAllPostByCategory(Long categoryId);
 	
 	//Get all Posts by User
-	List<PostDto> getAllPostByUser(Integer userId);
+	List<PostDto> getAllPostByUser(Long userId);
 	
 	//Searching sort using Keyword
 	List<PostDto> searchPosts(String keyword);
